requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
fireworks-ai==0.15.8
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
httpx-ws==0.6.2
idna==3.10
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.3
wsproto==1.2.0
